requirements.txt
discord.py
python-dotenv
sentence-transformers
numpy
scikit-learn
faiss-cpu
